// lib/api/types.gen.ts
export type MeasurementModel = {
    typeId?: number;
    reading?: number;
    takenAt?: string;
};

export type MeasurementTypeModel = {
    id?: number;
    name?: string | null;
    unit?: string | null;
    capacity?: number;
};